linking interface with child of container view
Buttons captured in game interface take affect in game screen even
though it is in another view.

left right work.
dash works.
shaking kinda works. meaning it is detected but cancles intermittently

jump is not connected.
rotate is not connected.
attack is not connected.

Most of today was figuring out the best way to approch the segways.
how to get to one and how to get back.  seems limited to one way in and
one way out.  If you want more a navagational controller would be
needed.